requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
qrcode==7.4.2
Pillow==10.3.0
python-dotenv==1.0.1
